Fixed the Makefile to make bin and obj
Turns out Git doesn't have any concept of directories and only knows
they exist based on their contents, which are ignored by the
`.gitignore`. I thought I was committing the empty dirs but I was not.

Anyway, the fix is to `mkdir -p` the object file directory in the object
file targets, and the binary directory in the final binary target.

Thanks to @rak1507 for pointing out this issue.